task-management-backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
mysql-connector-python==8.0.27

pydantic~=1.10.15
python-dotenv~=1.0.1
pymysql~=1.1.0
starlette~=0.14.2